split behavior.rs into decision/, patrol.rs + roadmap updates
