+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 2026-03-14
 
+- **Split behavior.rs** -- extracted 4318-line monolith into 4 files: `decision/mod.rs` (2800 lines: decision_system + transition helpers + utility AI + flee/leash), `decision/tests.rs` (1009 lines: 36 tests), `patrol.rs` (321 lines: on_duty_tick_system + rebuild_patrol_routes_system + 7 tests), `behavior.rs` coordinator (203 lines: SystemParam bundles + arrival_system). Re-exports preserve all external imports. 266 tests passing, zero behavior change.
+- **Roadmap updates** -- Stage 16 cache-friendly building iteration checked off (done in 03-11). Stage 26 marked 3 items complete (resource types, harvester buildings, resource storage). Current sprint updated to reflect active work. Code health backlog added (god-file splits for game_hud.rs, ai_player.rs). Test count updated to 261->266.
 - **NPC activity controller** -- refactored NPC behavior into deterministic `Activity.kind + Activity.phase + Activity.target` model. Added `ActivityPhase` (Ready/Transit/Active/Holding) and `ActivityTarget` (None/Home/Fountain/PatrolPost/SquadPoint/Worksite/RaidPoint/Dropoff/WanderPoint) enums. All 10 activity kinds use explicit phase lifecycle via `transition_activity()` / `transition_phase()` helpers -- zero raw `activity.kind =` writes remain. Removed vestigial `Activity.worksite` and `Activity.target_pos` fields. Added `reason: &'static str` and `last_frame: u32` debug fields for BRP inspection. Priority 0 arrival gate restricted to Transit/Ready phases -- fixes Rest/Heal infinite arrival trap bug. SquadAttack lifecycle separated from Patrol (distinct targets, distinct arrival handling). `on_duty_tick_system` phase-gated to Holding only. Farmer en-route retarget gated on Transit phase. Save/load maps phase/target for all activities. BRP output includes activity_phase, activity_target, transition_reason, last_transition_frame. 223 cargo tests (30 new lifecycle/transition/phase tests). Breaks old saves (intentional per spec).
 - **Save round-trip fixes** -- Work, Mine, SquadAttack now preserve phase+target through save/load. Added `GoingToWork`, `GoingToSquadTarget` save variants. 12 round-trip tests covering all activity kinds including transit states.
 - **Homeless NPC fix** -- NPCs whose home building is destroyed now get Home reassigned to their town's fountain position instead of Home(-1,-1). All `home != Vec2::ZERO` checks replaced with `home_valid` derived from `Home::is_valid()`. `transition_to_rest()` helper consolidates fountain fallback logic. 5 tests including full decision_system integration test.
