+++ rust/src/systems/mod.rs
@@ -2,8 +2,9 @@
 
 pub mod ai_player;
 pub mod audio;
-mod behavior;
+pub(crate) mod behavior;
 mod combat;
+mod decision;
 mod drain;
 mod economy;
 mod energy;
@@ -11,6 +12,7 @@ mod health;
 pub mod llm_player;
 mod movement;
 pub mod pathfinding;
+mod patrol;
 pub mod remote;
 pub(crate) mod spawn;
 pub mod stats;
@@ -21,11 +23,13 @@ pub use ai_player::{
 };
 pub use behavior::*;
 pub use combat::*;
+pub use decision::decision_system;
 pub use drain::*;
 pub use economy::*;
 pub use energy::*;
 pub use health::*;
 pub use movement::*;
+pub use patrol::{on_duty_tick_system, rebuild_patrol_routes_system};
 pub use spawn::*;
 pub use stats::{
     CombatConfig, UPGRADES, UpgradeMsg, auto_upgrade_system, expansion_cost, level_from_xp,
